fix: resolve remaining admin review issues
